refactor: change general dir structure
chore: install typescript
