+++ src/app.ts
@@ -1,21 +1,21 @@
-var express = require('express');
-var bodyParser = require('body-parser');
-var app = express();
 
-var teste1 = require("./teste1");
-var teste2 = require("./teste2");
-var teste3 = require("./teste3");
-var teste4 = require("./teste4");
-var teste5 = require("./teste5");
+const express = require('express');
+const app = express();
+
+import routes from "./routes";
+
+const teste1 = require("./challenges/teste1");
+const teste2 = require("./challenges/teste2");
+const teste3 = require("./challenges/teste3");
+const teste4 = require("./challenges/teste4");
+const teste5 = require("./challenges/teste5");
 
 
 app.set('view engine', 'jade');
 
 app.use(express.json());
 app.use(express.urlencoded());
-
-app.use(bodyParser.json());                        
-app.use(bodyParser.urlencoded({ extended: true }));
+app.use(routes)
 
 app.use(express.static(__dirname + '/public'));
 
@@ -36,7 +36,5 @@ app.put("/users", teste4)
 app.get("/users/access", teste5);
 
 
-const port  = 3000;
-app.listen(port, function(){
-  console.log('Express server listening on port ' + port);
-});
+const port  = 8080;
+app.listen({ port, host: "0.0.0.0" }, () => console.log("Server is running smoothly on port:" + port));
