+++ src/challenges/teste1.js
@@ -1,4 +1,4 @@
-var data =  require("./fakeData");
+var data =  require("../api/fakeData");
 
 const getUser = ( req, res, next ) => {
